fix: swagger documentation in controllers
